Make UniversalEventsToken optional
Rationale: Currently when making a very simple Fusion.js application with only a router, you're also required to pull in the UniversalEventsToken, which also requires FetchToken. It feels quite complex for a small non-Uber application, and making this token optional will do quite a lot to ease the first-time app developer.